+++ activerecord/lib/active_record/connection_adapters/abstract/schema_definitions.rb
@@ -186,6 +186,10 @@ module ActiveRecord
       end
       alias validated? validate?
 
+      def enforced?
+        options.fetch(:enforced, true)
+      end
+
       def export_name_on_schema_dump?
         !ActiveRecord::SchemaDumper.chk_ignore_pattern.match?(name) if name
       end